changed: split legacy VFP test into own file
